Fix execute_command test
The ExecuteCommand DSDL got an "output" field in version 1.3. The unittest
in yakut is failing, because the test does not account for this field.
This commit changes the test such that it treats the "output" field as
optional, i.e. the test passes with or without the field present in the output
to preserve backwards compatibility with prior DSDL versions.